pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jdass"
version = "0.1.0"
description = "Job Discovery and Application System — local-first, LLM-powered"
requires-python = ">=3.12"
dependencies = [
    # Storage
    "sqlmodel>=0.0.21",
    "aiosqlite>=0.20.0",

    # HTTP / scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.44.0",

    # Config / env
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",

    # Scheduling
    "apscheduler>=3.10.4",

    # Logging
    "loguru>=0.7.2",

    # Resume generation
    "python-docx>=1.1.2",
    "weasyprint>=62.3",

    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",

    # Dashboard
    "streamlit>=1.36.0",
    "pandas>=2.2.0",
    "plotly>=5.22.0",

    # Utilities
    "tenacity>=8.4.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
]

[project.scripts]
jdass-discover   = "pipelines.discovery:main"
jdass-dashboard  = "dashboard.ui:main"
jdass-scheduler  = "scheduler.jobs:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = [
    "scrapers", "core", "outreach", "automation",
    "dashboard", "storage", "llm", "pipelines", "scheduler",
]
